create abstract location class
